Struct spirv::ImageOperands

pub struct ImageOperands { /* private fields */ }

SPIR-V operand kind: ImageOperands

Implementations

impl ImageOperands

pub const NONE: ImageOperands

pub const BIAS: ImageOperands

pub const LOD: ImageOperands

pub const GRAD: ImageOperands

pub const CONST_OFFSET: ImageOperands

pub const OFFSET: ImageOperands

pub const CONST_OFFSETS: ImageOperands

pub const SAMPLE: ImageOperands

pub const MIN_LOD: ImageOperands

pub const MAKE_TEXEL_AVAILABLE: ImageOperands

pub const MAKE_TEXEL_AVAILABLE_KHR: ImageOperands

pub const MAKE_TEXEL_VISIBLE: ImageOperands

pub const MAKE_TEXEL_VISIBLE_KHR: ImageOperands

pub const NON_PRIVATE_TEXEL: ImageOperands

pub const NON_PRIVATE_TEXEL_KHR: ImageOperands

pub const VOLATILE_TEXEL: ImageOperands

pub const VOLATILE_TEXEL_KHR: ImageOperands

pub const SIGN_EXTEND: ImageOperands

pub const ZERO_EXTEND: ImageOperands

pub const fn empty() -> ImageOperands

Returns an empty set of flags

pub const fn all() -> ImageOperands

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u32) -> Option<ImageOperands>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> ImageOperands

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u32) -> ImageOperands

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: ImageOperands) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: ImageOperands) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: ImageOperands)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: ImageOperands)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: ImageOperands)

Toggles the specified flags in-place.

pub fn set(&mut self, other: ImageOperands, value: bool)

Inserts or removes the specified flags depending on the passed value.