Struct spirv::MemoryAccess

pub struct MemoryAccess { /* fields omitted */ }

SPIR-V operand kind: MemoryAccess

Implementations

impl MemoryAccess

pub const NONE: MemoryAccess

pub const VOLATILE: MemoryAccess

pub const ALIGNED: MemoryAccess

pub const NONTEMPORAL: MemoryAccess

pub const MAKE_POINTER_AVAILABLE: MemoryAccess

pub const MAKE_POINTER_AVAILABLE_KHR: MemoryAccess

pub const MAKE_POINTER_VISIBLE: MemoryAccess

pub const MAKE_POINTER_VISIBLE_KHR: MemoryAccess

pub const NON_PRIVATE_POINTER: MemoryAccess

pub const NON_PRIVATE_POINTER_KHR: MemoryAccess

pub const fn empty() -> MemoryAccess

Returns an empty set of flags

pub const fn all() -> MemoryAccess

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u32) -> Option<MemoryAccess>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> MemoryAccess

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u32) -> MemoryAccess

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: MemoryAccess) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: MemoryAccess) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: MemoryAccess)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: MemoryAccess)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: MemoryAccess)

Toggles the specified flags in-place.

pub fn set(&mut self, other: MemoryAccess, value: bool)

Inserts or removes the specified flags depending on the passed value.